[CSL-2212] Eliminate `forkFinally` in `launchNotifier`
Previously it was using `forkForever`, now threads are spawned using
`mapConcurrently` which is generally safer.
The logic is quite similar anyway. But there is a subtle difference: new
version doesn't catch async exceptions from notifier. They will be propagated
to the higher level and I think that's exactly what we want.